+++ R/find_parameters.R
@@ -177,7 +177,7 @@ find_parameters.brmultinom <- find_parameters.multinom
 find_parameters.multinom_weightit <- function(x, flatten = FALSE, ...) {
   params <- stats::coef(x)
   resp <- gsub("(.*)~(.*)", "\\1", names(params))
-  pars <- gsub("(.*)~(.*)", "\\2", names(params))[resp == resp[1]]
+  pars <- list(conditional = gsub("(.*)~(.*)", "\\2", names(params))[resp == resp[1]])
 
   if (flatten) {
     unique(unlist(pars, use.names = FALSE))