+++ R/link_function.R
@@ -36,28 +36,37 @@ link_function.default <- function(x, ...) {
     x <- x$gam
     class(x) <- c(class(x), c("glm", "lm"))
   }
+  .extract_generic_linkfun(x)
+}
 
-  tryCatch(
-    {
-      # get model family
-      ff <- .gam_family(x)
 
+.extract_generic_linkfun <- function(x, default_link = NULL) {
+  # general approach
+  out <- .safe(stats::family(x)$linkfun)
+  # if it fails, try to retrieve from model information
+  if (is.null(out)) {
+    # get model family, consider special gam-case
+    ff <- .gam_family(x)
+    if ("linkfun" %in% names(ff)) {
       # return link function, if exists
-      if ("linkfun" %in% names(ff)) {
-        return(ff$linkfun)
-      }
-
+      out <- ff$linkfun
+    } else if ("link" %in% names(ff) && is.character(ff$link)) {
       # else, create link function from link-string
-      if ("link" %in% names(ff)) {
-        return(match.fun(ff$link))
+      out <- .safe(stats::make.link(link = ff$link)$linkfun)
+      # or match the function - for "exp()", make.link() won't work
+      if (is.null(out)) {
+        out <- .safe(match.fun(ff$link))
       }
-
-      NULL
-    },
-    error = function(x) {
-      NULL
     }
-  )
+  }
+  # if all fails, force default link
+  if (is.null(out) && !is.null(default_link)) {
+    out <- switch(default_link,
+      identity = .safe(stats::gaussian(link = "identity")$linkfun),
+      .safe(stats::make.link(link = default_link)$linkfun)
+    )
+  }
+  out
 }
 
 
@@ -66,7 +75,7 @@ link_function.default <- function(x, ...) {
 
 #' @export
 link_function.lm <- function(x, ...) {
-  stats::gaussian(link = "identity")$linkfun
+  .extract_generic_linkfun(x, "identity")
 }
 
 #' @export
@@ -202,7 +211,7 @@ link_function.nestedLogit <- function(x, ...) {
 
 #' @export
 link_function.multinom <- function(x, ...) {
-  stats::make.link(link = "logit")$linkfun
+  .extract_generic_linkfun(x, "logit")
 }
 
 #' @export
@@ -481,7 +490,7 @@ link_function.cglm <- function(x, ...) {
   method <- parse(text = safe_deparse(x$call))[[1]]$method
 
   if (!is.null(method) && method == "clm") {
-    link <- "identiy"
+    link <- "identity"
   }
   stats::make.link(link = link)$linkfun
 }
@@ -544,28 +553,7 @@ link_function.bcplm <- link_function.cpglmm
 
 #' @export
 link_function.gam <- function(x, ...) {
-  lf <- tryCatch(
-    {
-      # get model family
-      ff <- .gam_family(x)
-
-      # return link function, if exists
-      if ("linkfun" %in% names(ff)) {
-        return(ff$linkfun)
-      }
-
-      # else, create link function from link-string
-      if ("link" %in% names(ff)) {
-        return(match.fun(ff$link))
-      }
-
-      NULL
-    },
-    error = function(x) {
-      NULL
-    }
-  )
-
+  lf <- .extract_generic_linkfun(x)
   if (is.null(lf)) {
     mi <- .gam_family(x)
     if (object_has_names(mi, "linfo")) {